Update SubModule.xml
Loading this mod before MCM causes a bug in the mod option screen. This is a dirty fix to the problem but there has to be a better way to fix this.

It might help to add MCM as a Nexus requirement, even though it is packaged into the mod, so that future versions of MCM will still work with this mod.