Patch skins.xml loader
The game combines skins.xml documents and seems to use the first `skins` array element (or the first matching `skin` nodes; implementation unclear) in the combined document for FaceGen. This means the extra sliders only show up if they are in a mod that is loaded before Native.

Fixing this involves patching the skins.xml loader to reverse the file loading order, so modules loaded later will load their versions of skins.xml first. This will undoubtably break existing skins.xml mods, but the alternative is to rename DCC so that it doesn't break MCM, and the latter looks outright ugly by comparison.

An alternative would be instead to modify the code so that Native's skins.xml is loaded last. This would maintain (a degree of) compatibility with skins.xml mods while guaranteeing that Native won't override modded skins.xml files.

Also fixed post-build directives because a folder is only a folder if it ends with a back slash.