Clean up the encyclopedia pages
Relocated the buttons to the right side of the encyclopedia to make them feel more uniform.

Also added in mod configuration setting to enable or disable the label for "DCC Options" on the encyclopedia. It's enabled by default, but can be disabled for experienced users who are aware of which buttons DCC adds, to keep the page looking clean.