Button placement based on defined aspect ratio
I've created a list of common aspect ratios used by the game and defined offsets for it. I suppose theoretically this could be done with a formula, but the range for aspect ratios is fairly straightforward, so this should be more than enough.

It is quick and easy to add new ranges for any aspect ratios that I do not have access to; such as 21:9.

This should cover most of them, however.

Also added (and kept) a debug command to return the aspect ratio, in-case we need it for future use-cases.

The aspect library could technically be tweaked and modified for other use-cases, but at this time it was not necessary, so I've left it as a primary caretaker for this function. Feel free to adjust it as you see fit.