Rework of the aspect ratio logic
This isn't entirely perfect, but it's accurate enough that it should suffice for our needs.

It's at least (I think) a better approach than having a custom margin offset library that may require custom tweaking for non-native resolutions, such as people manually resizing the game window.

At least this way it should dynamically fit regardless of the aspect ratio of the game client.