Separate character screen code to make it work
Turns out if Bannerlord didn't register the screen for the game state, then trying to push the game state will crash the game because the screen is missing.

This complicated stuff needs to be done.

Will do campaign behavior separation next. Dependency injection will be used, like the other separations.